Add event contracts module and update inventory and order services to use shared DTOs
