+++ pom.xml
@@ -20,6 +20,7 @@
     <description>Monorepo for Microservices</description>
 
     <modules>
+        <module>services/event-contracts</module>
         <module>services/product-service</module>
         <module>services/inventory-service</module>
         <module>services/review-service</module>
