+++ services/event-contracts/src/main/java/nik/kalomiris/events/dtos/OrderLineItem.java
@@ -1,0 +1,30 @@
+package nik.kalomiris.events.dtos;
+
+public class OrderLineItem {
+    private Long productId;
+    private Integer quantity;
+
+    public OrderLineItem() {
+    }
+
+    public OrderLineItem(Long productId, Integer quantity) {
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Long productId) {
+        this.productId = productId;
+    }
+
+    public Integer getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(Integer quantity) {
+        this.quantity = quantity;
+    }
+}
