+++ services/event-contracts/src/main/java/nik/kalomiris/events/dtos/ProductCreatedEvent.java
@@ -1,0 +1,19 @@
+package nik.kalomiris.events.dtos;
+
+public class ProductCreatedEvent {
+    private String sku;
+
+    public ProductCreatedEvent() {}
+
+    public ProductCreatedEvent(String sku) {
+        this.sku = sku;
+    }
+
+    public String getSku() {
+        return sku;
+    }
+
+    public void setSku(String sku) {
+        this.sku = sku;
+    }
+}
