+++ services/inventory-service/src/main/java/nik/kalomiris/inventory_service/events/dtos/OrderEvent.java
@@ -1,52 +1,13 @@
 package nik.kalomiris.inventory_service.events.dtos;
 
-import java.util.List;
-import java.time.Instant;
-
+/**
+ * Deprecated local DTO.
+ *
+ * @deprecated Use {@link nik.kalomiris.events.dtos.OrderEvent} from the shared
+ * `event-contracts` module instead.
+ */
+@Deprecated(since = "2025-10-27", forRemoval = true)
 public class OrderEvent {
-    private String orderNumber;
-    private String correlationId;
-    private Instant timestamp;
-    private List<OrderLineItem> lineItems;
-
-    public OrderEvent() {}
-
-    public OrderEvent(String orderNumber, String correlationId, Instant timestamp, List<OrderLineItem> lineItems) {
-        this.orderNumber = orderNumber;
-        this.correlationId = correlationId;
-        this.timestamp = timestamp;
-        this.lineItems = lineItems;
-    }
-
-    public String getOrderNumber() {
-        return orderNumber;
-    }
-
-    public void setOrderNumber(String orderNumber) {
-        this.orderNumber = orderNumber;
-    }
-
-    public String getCorrelationId() {
-        return correlationId;
-    }
-
-    public void setCorrelationId(String correlationId) {
-        this.correlationId = correlationId;
-    }
-
-    public Instant getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(Instant timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public List<OrderLineItem> getLineItems() {
-        return lineItems;
-    }
-
-    public void setLineItems(List<OrderLineItem> lineItems) {
-        this.lineItems = lineItems;
-    }
+    // Deprecated shim - intentionally empty. Migrate to
+    // nik.kalomiris.events.dtos.OrderEvent and remove this class.
 }
