+++ services/inventory-service/src/main/java/nik/kalomiris/inventory_service/events/dtos/ProductCreatedEvent.java
@@ -1,20 +1,13 @@
 package nik.kalomiris.inventory_service.events.dtos;
 
+/**
+ * Deprecated local DTO.
+ *
+ * @deprecated Use {@link nik.kalomiris.events.dtos.ProductCreatedEvent} from the
+ * shared `event-contracts` module instead.
+ */
+@Deprecated(since = "2025-10-27", forRemoval = true)
 public class ProductCreatedEvent {
-    private String sku;
-
-    public ProductCreatedEvent() {
-    }
-
-    public ProductCreatedEvent(String sku) {
-        this.sku = sku;
-    }
-
-    public String getSku() {
-        return sku;
-    }
-
-    public void setSku(String sku) {
-        this.sku = sku;
-    }
+    // Deprecated shim - intentionally empty. Migrate to
+    // nik.kalomiris.events.dtos.ProductCreatedEvent and remove this class.
 }
