+++ services/order-service/src/main/java/nik/kalomiris/order_service/service/OrderService.java
@@ -9,11 +9,13 @@ import nik.kalomiris.logging_client.LogMessage;
 import nik.kalomiris.order_service.config.RabbitMQConfig;
 import nik.kalomiris.order_service.domain.Order;
 import nik.kalomiris.order_service.domain.OrderLineItem;
-import nik.kalomiris.order_service.dto.OrderPlacedEvent;
+import nik.kalomiris.events.dtos.OrderEvent;
 import nik.kalomiris.order_service.dto.OrderRequest;
 import nik.kalomiris.order_service.mapper.OrderMapper;
 import nik.kalomiris.order_service.repository.OrderRepository;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.transaction.support.TransactionSynchronization;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -60,16 +62,29 @@ public class OrderService {
 
         orderRepository.save(order);
 
-        OrderPlacedEvent event = new OrderPlacedEvent(
+        OrderEvent event = new OrderEvent(
             order.getOrderNumber(),
             order.getOrderNumber(),
             Instant.now(),
             order.getOrderLineItems()
                 .stream()
-                .map(li -> new OrderPlacedEvent.LineItem(li.getProductId(), li.getQuantity()))
+                .map(li -> new nik.kalomiris.events.dtos.OrderLineItem(li.getProductId(), li.getQuantity()))
                 .toList()
         );
-        rabbitTemplate.convertAndSend(RabbitMQConfig.EXCHANGE_NAME, RabbitMQConfig.ROUTING_KEY_ORDER_CREATED, event);
+
+        // Ensure we publish the event only after the database transaction commits so
+        // consumers won't receive the event before the order is visible in the DB.
+        if (TransactionSynchronizationManager.isSynchronizationActive()) {
+            TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
+                @Override
+                public void afterCommit() {
+                    rabbitTemplate.convertAndSend(RabbitMQConfig.EXCHANGE_NAME, RabbitMQConfig.ROUTING_KEY_ORDER_CREATED, event);
+                }
+            });
+        } else {
+            // No transaction active (e.g., tests or manual call) — send immediately.
+            rabbitTemplate.convertAndSend(RabbitMQConfig.EXCHANGE_NAME, RabbitMQConfig.ROUTING_KEY_ORDER_CREATED, event);
+        }
 
         // Publish a log event about the order creation. Ignore logging failures.
         try {
