+++ services/product-service/pom.xml
@@ -66,6 +66,12 @@
 			<artifactId>logging-client</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
+		<!-- Shared event contracts -->
+		<dependency>
+			<groupId>nik.kalomiris</groupId>
+			<artifactId>event-contracts</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
 
 	<build>
